Moved query and result sections in the separated component
